// src/interfaces/lottie.ts
export enum BlendMode {
    Normal,
    Multiply,
    Screen,
    Overlay,
    Darken,
    Lighten,
    ColorDodge,
    ColorBurn,
    HardLight,
    SoftLight,
    Difference,
    Exclusion,
    Hue,
    Saturation,
    Color,
    Luminosity,
  }
  
  export enum MatteMode {
    Normal,
    Alpha,
    InvertedAlpha,
    Luma,
    InvertedLuma,
  }
  
  export enum Layer3DMode {
    Off,
    On,
  }
  
  export enum AutoOrientMode {
    Off,
    On,
  }
// ...
  export type Effect = {
    /** Effect Index. */
    ix: number;
    /** After Effect's Name. */
    nm: string;
    /** Effect type. */
    ty: EffectType;
    /** Effect parameters. [] */
    ef: EffectValue[];
    /** ??? */
    np: number;
    /** Match name (for expressions) */
    mn: string;
    /** ??? */
    en: number;
  };
  
  export const enum MaskMode {
    No = 'n',
    Add = 'a',
    Subtract = 's',
    Intersect = 'i',
    Lighten = 'l',
    Darken = 'd',
    Difference = 'f',
  }
  
  export type Mask = {
    /** Inverted Mask flag. */
    inv: boolean;
    /** Mask name. */
    nm?: string;
    /** Mask vertices. */
    pt: ShapeProperty;
    /** Mask opacity. */
    o: Value;
    /** Mask mode. */
    mode: MaskMode;
    /** Dilate. */
    x: Value;
  };
  
  export type Layer = {
    /** 3d layer flag */
    ddd?: Layer3DMode;
    /** Hidden layer */
    hd?: boolean;
    /** Layer type */
    ty: LayerType;
    /** After Effects Layer Name */
    nm?: string;
    /** Layer Parent */
    parent?: number;
    /** Layer Time Stretching */
    sr?: number;
    /** Transform properties */
    ks: Transform;
    /** Auto-Orient along path AE property */
    ao?: AutoOrientMode;
    /** In Point of layer */
    ip: number;
    /** Out Point of layer */
    op: number;
    /** Start Time of layer */
    st: number;
    /** Blend Mode */
    bm?: BlendMode;
    /** Matte mode, the layer will inherit the transparency from the layer above */
    tt?: MatteMode;
    /** Layer index in AE */
    ind?: number;
    /** Whether the layer has some masks applied */
    hasMask?: boolean;
    /** List of Masks */
    masksProperties?: Mask[];
    /** List of Effects */
    ef?: Effect[];
    /** matte_target */
    td?: number;
    /**
     * Enable motion blur
     * (only has effect if Animation.mb is set)
     */
    mb?: boolean;
  };
  
  export enum LayerType {
    precomp,
    solid,
    still,
    null,
    shape,
    text,
    audio,
    pholderVideo,
    imageSeq,
    video,
    pholderStill,
    guide,
    adjustment,
    camera,
    light,
  }
  
  export type Transform = {
    /** Transform Anchor Point. */
    a: MultiDimensional;
    /** Transform Position. */
    p: PositionValue;
    /** Transform Scale. */
    s: MultiDimensional;
    /** Transform Rotation. */
    r: Value;
    /** Transform Opacity. */
    o?: Value;
    /** Transform Skew. */
    sk?: Value;
    /** Transform Skew Axis. */
    sa?: Value;
  };
// ...
  export type ShapeLayer = Layer & {
    /** Shape list of items. */
    shapes: ShapeElement[];
  };
// ...
  export enum ShapeType {
    Group = 'gr',
    Rect = 'rc',
    Ellipse = 'el',
    Fill = 'fl',
    TransformShape = 'tr',
    Path = 'sh',
    Stroke = 'st',
  }
  
  export type ShapeElement = {
    /** Hide element. */
    hd: boolean;
  
    /** After Effect's Name. */
    nm: string;
  
    /** Shape type. */
    ty: ShapeType;
  
    /** Property index. */
    cix: number;
  
    /** Blending Mode */
    bm: BlendMode;
  };
  
  export type GroupShapeElement = ShapeElement & {
    /** Number of properties. */
    np: number;
    /** List of ShapeElement.  */
    it: Array<ShapeElement>;
  };
  
  export type Bezier = {
    /** Closed property of shape. */
    c?: boolean;
    /** Cubic bezier handles for the segments before each vertex. */
    i: number[][];
    /** Cubic bezier handles for the segments after each vertex. */
    o: number[][];
    /** Bezier curve vertices. */
    v: number[][];
  };
  
  export type ShapePropKeyframe = {
    /** Start time of keyframe segment. */
    t: number;
    /** Bezier curve easing in value. */
    i: KeyframeBezierHandle;
    /** Bezier curve easing out value. */
    o: KeyframeBezierHandle;
    /** [0-1] Jump to the end value. */
    h: number;
    /** or list of Bezier	Start value of keyframe segment. */
    s: Bezier;
    /** or list of Bezier	End value of keyframe segment. */
    e: Bezier;
  };
  
  export type ShapeProperty = {
    /**
     * Bezier: Non-animated value.
     * ShapePropKeyframe[]: Keyframe list.
     */
    k: Bezier | ShapePropKeyframe[];
    /** Property index. */
    ix: number;
    /** [0-1] Whether it's animated. */
    a?: number;
  };
  
  export type PathShape = ShapeElement & {
    /** Shape's vertices. */
    ks: ShapeProperty;
    /** index  */
    ind: number;
  };
  
  export type MultiDimensional = {
    /**
     * number[]: Non-animated value.
     * OffsetKeyframe[]: Animated keyframes.
     */
    k: number[] | OffsetKeyframe[];
    /** Property index. */
    ix?: number;
    /** [0-1] Whether it's animated. */
    a?: number;
    /** ??? */
    l?: number;
  };
  
  export type Value = {
    /**
     * number | number[]: Non-animated value.
     * OffsetKeyframe[]: Animated keyframes.
     */
    k: number | number[] | OffsetKeyframe[];
    /** Property index. */
    ix?: number;
    /** [0-1] Whether it's animated. */
    a?: number;
    /** Split values (???) */
    s?: boolean;
  };
  
  export type PositionValue = {
    /**
     * number | number[]: Non-animated value.
     * OffsetKeyframe[]: Animated keyframes.
     */
    k: number | number[] | OffsetKeyframe[];
    /** Property index. */
    ix?: number;
    /** [0-1] Whether it's animated. */
    a?: number;
    /** ??? */
    l?: number;
  };
  
  export type KeyframeBezierHandle = {
    /** x position of the handle. */
    x: number | number[];
    /** y position of the handle. */
    y: number | number[];
  };
  
  export type OffsetKeyframe = {
    /** Start time of keyframe segment. */
    t: number;
    /** Bezier curve easing in value. */
    i?: KeyframeBezierHandle;
    /** Bezier curve easing out value. */
    o?: KeyframeBezierHandle;
    /** [0-1] Jump to value. */
    h?: number;
    /** Start value of keyframe segment. */
    s: number | number[];
    /** End value of keyframe segment. */
    e?: number[];
    /** In Spatial Tangent. Only for spatial properties. */
    ti?: number[];
    /** Out Spatial Tangent. Only for spatial properties. */
    to?: number[];
  };
  
  export type RectShape = ShapeElement & {
    /** After Effect's Direction. */
    d: number;
    /** Rect's position. */
    p: MultiDimensional;
    /** Rect's size. */
    s: MultiDimensional;
    /** Rect's rounded corners. */
    r: Value;
  };
  
  export type TransformShape = ShapeElement & {
    /** Anchor point */
    a: MultiDimensional;
    /** Transform Position. */
    p: PositionValue;
    /** Transform Scale. */
    s: MultiDimensional;
    /** Transform Rotation. */
    r: Value;
    /** Transform Opacity. */
    o: Value;
    /** Transform Skew. */
    sk: Value;
    /** Transform Skew Axis. */
    sa: Value;
  };
  
  export type Color4 = [number, number, number, number];
  
  export type ColorValue = {
    /**
     * number[r, g, b]: Non-animated value.
     * OffsetKeyframe[]: Animated keyframes.
     */
    k: [number, number, number] | OffsetKeyframe[];
    /** Property index. */
    ix: number;
    /** [0-1] Whether it's animated. */
    a?: number;
  };
  
  export enum FillRule {
    NonZero = 1,
    EvenOdd,
  }
  
  export type FillShape = ShapeElement & {
    /** Fill Opacity. */
    o: Value;
    /** Fill Color. */
    c: ColorValue;
    /** Fill rule. */
    r: FillRule;
  };